% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry.R
\name{analyze}
\alias{analyze}
\title{analyze}
\usage{
analyze(input, output, session, M, D, uid)
}
\arguments{
\item{input}{is shiny input variable}

\item{output}{is shiny output variable}

\item{session}{is shiny session variable}

\item{M}{is the meta data connection structure}

\item{D}{is the data connection structure}

\item{uid}{is the logged in user id}

\item{id}{is the caller id}
}
\description{
the query module
}
