% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash.R
\name{dashboard}
\alias{dashboard}
\title{dashboardUI}
\usage{
dashboard(input, output, session, dashid, M)
}
\arguments{
\item{input}{is shiny input variable}

\item{output}{is shiny output variable}

\item{session}{is shiny session variable}

\item{dashid}{is the folder id of reports}

\item{M}{is the meta data connection structure}
}
\description{
UI to display a dashboard
}
