% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash.R
\name{dashboards}
\alias{dashboards}
\title{reportscharts}
\usage{
dashboards(input, output, session, M)
}
\arguments{
\item{input}{is shiny input variable}

\item{output}{is shiny output variable}

\item{session}{is shiny session variable}

\item{M}{is the meta data connection structure}
}
\description{
Display saved reports
}
