% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash.R
\name{dashboardUI}
\alias{dashboardUI}
\title{dashboardUI}
\usage{
dashboardUI(id, dashid, dashname, M)
}
\arguments{
\item{id}{is caller id}

\item{dashid}{is the folder id of reports}

\item{dashname}{is the name of the folder}

\item{M}{is the meta data connection structure}
}
\description{
UI to display a dashboard
}
