% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reps.R
\name{reportscharts}
\alias{reportscharts}
\title{reportscharts}
\usage{
reportscharts(input, output, M, D, uid)
}
\arguments{
\item{input}{is shiny input variable}

\item{output}{is shiny output variable}

\item{M}{is the meta data connection structure}

\item{D}{is the data connection structure}

\item{uid}{is the logged in user id}
}
\description{
Display saved reports
}
