% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry.R
\docType{data}
\name{rg}
\alias{rg}
\title{rg}
\format{An object of class \code{reactivevalues} of length 3.}
\usage{
rg
}
\description{
reactive values for the 'graph': containing data from SQL query and chart properties
}
\keyword{datasets}
