% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry_save.R
\name{savereport}
\alias{savereport}
\title{savereport}
\usage{
savereport(input, output, session, g, M, uid, currfilt)
}
\arguments{
\item{input}{is shiny input variable}

\item{output}{is shiny output variable}

\item{session}{is shiny session variable}

\item{g}{is graph structure to be saved as a report in database}

\item{M}{is the meta data connection structure}

\item{uid}{is the logged in user id}

\item{currfilt}{contains the filters applied for this report}
}
\description{
save the query as a report
}
