% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry_save.R
\name{savereportUI}
\alias{savereportUI}
\title{savereportUI}
\usage{
savereportUI(id, ft)
}
\arguments{
\item{id}{is the caller id}

\item{ft}{is the folder to save the report in}
}
\description{
UI to save the query as a report
}
