% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry.R
\docType{data}
\name{xr}
\alias{xr}
\title{modqry}
\format{An object of class \code{reactivevalues} of length 3.}
\usage{
xr
}
\description{
A 'shiny' Module for Creating SQL queries via a front end

reactive values for this module ('xplorer')
}
\keyword{datasets}
